+++ tests/test_backward_layers.py
@@ -11,14 +11,6 @@ from decomon.backward_layers.backward_layers import get_backward
 from decomon.layers.decomon_layers import DecomonActivation, DecomonFlatten
 from decomon.layers.decomon_reshape import DecomonReshape
 
-from . import (
-    assert_output_properties_box_linear,
-    get_standard_values_multid_box,
-    get_standart_values_1d_box,
-    get_tensor_decomposition_1d_box,
-    get_tensor_decomposition_multid_box,
-)
-
 # Activation
 
 
@@ -195,7 +187,7 @@ from . import (
         (6, None, "ibp", True, 32),
     ],
 )
-def test_Backward_Activation_1D_box_model(n, activation, mode, previous, floatx):
+def test_Backward_Activation_1D_box_model(n, activation, mode, previous, floatx, helpers):
     K.set_floatx("float{}".format(floatx))
     eps = K.epsilon()
     decimal = 5
@@ -205,8 +197,8 @@ def test_Backward_Activation_1D_box_model(n, activation, mode, previous, floatx)
 
     layer = DecomonActivation(activation, dc_decomp=False, mode=mode, dtype=K.floatx())
 
-    inputs = get_tensor_decomposition_1d_box(dc_decomp=False)
-    inputs_ = get_standart_values_1d_box(n, dc_decomp=False)
+    inputs = helpers.get_tensor_decomposition_1d_box(dc_decomp=False)
+    inputs_ = helpers.get_standart_values_1d_box(n, dc_decomp=False)
     x, y, z_, u_c, W_u, b_u, l_c, W_l, b_l = inputs_
 
     if mode == "hybrid":
@@ -246,7 +238,7 @@ def test_Backward_Activation_1D_box_model(n, activation, mode, previous, floatx)
         model = Model(inputs[2:], [w_out_u_, b_out_u_, w_out_l_, b_out_l_])
         w_u_, b_u_, w_l_, b_l_ = model.predict(inputs_[2:])
 
-    assert_output_properties_box_linear(
+    helpers.assert_output_properties_box_linear(
         x,
         None,
         z_[:, 0],
@@ -360,7 +352,7 @@ def test_Backward_Activation_1D_box_model(n, activation, mode, previous, floatx)
         (0, "relu", 16, "ibp", False),
     ],
 )
-def test_Backward_Activation_multiD_box(odd, activation, floatx, mode, previous):
+def test_Backward_Activation_multiD_box(odd, activation, floatx, mode, previous, helpers):
 
     K.set_floatx("float{}".format(floatx))
     eps = K.epsilon()
@@ -370,8 +362,8 @@ def test_Backward_Activation_multiD_box(odd, activation, floatx, mode, previous)
         decimal = 2
 
     layer = DecomonActivation(activation, dc_decomp=False, mode=mode, dtype=K.floatx())
-    inputs = get_tensor_decomposition_multid_box(odd, dc_decomp=False)
-    inputs_ = get_standard_values_multid_box(odd, dc_decomp=False)
+    inputs = helpers.get_tensor_decomposition_multid_box(odd, dc_decomp=False)
+    inputs_ = helpers.get_standard_values_multid_box(odd, dc_decomp=False)
     x, y, z_, u_c, W_u, b_u, l_c, W_l, b_l = inputs_
 
     if mode == "hybrid":
@@ -416,7 +408,7 @@ def test_Backward_Activation_multiD_box(odd, activation, floatx, mode, previous)
         output_ = f_dense(inputs_)
     w_u_, b_u_, w_l_, b_l_ = output_
 
-    assert_output_properties_box_linear(
+    helpers.assert_output_properties_box_linear(
         x,
         None,
         z_[:, 0],
@@ -530,7 +522,7 @@ def test_Backward_Activation_multiD_box(odd, activation, floatx, mode, previous)
         (1, 16, "ibp", False, "channels_first"),
     ],
 )
-def test_Backward_Flatten_multiD_box(odd, floatx, mode, previous, data_format):
+def test_Backward_Flatten_multiD_box(odd, floatx, mode, previous, data_format, helpers):
 
     if data_format == "channels_first" and not len(K._get_available_gpus()):
         return
@@ -543,8 +535,8 @@ def test_Backward_Flatten_multiD_box(odd, floatx, mode, previous, data_format):
         decimal = 2
 
     layer = DecomonFlatten("channels_last", dc_decomp=False, mode=mode, dtype=K.floatx())
-    inputs = get_tensor_decomposition_multid_box(odd, dc_decomp=False)
-    inputs_ = get_standard_values_multid_box(odd, dc_decomp=False)
+    inputs = helpers.get_tensor_decomposition_multid_box(odd, dc_decomp=False)
+    inputs_ = helpers.get_standard_values_multid_box(odd, dc_decomp=False)
     x, y, z_, u_c, W_u, b_u, l_c, W_l, b_l = inputs_
 
     if mode == "hybrid":
@@ -589,7 +581,7 @@ def test_Backward_Flatten_multiD_box(odd, floatx, mode, previous, data_format):
         output_ = f_dense(inputs_)
     w_u_, b_u_, w_l_, b_l_ = output_
 
-    assert_output_properties_box_linear(
+    helpers.assert_output_properties_box_linear(
         x,
         None,
         z_[:, 0],
@@ -684,7 +676,7 @@ def test_Backward_Flatten_multiD_box(odd, floatx, mode, previous, data_format):
         (1, 16, "ibp", False, "channels_first"),
     ],
 )
-def test_Backward_Reshape_multiD_box(odd, floatx, mode, previous, data_format):
+def test_Backward_Reshape_multiD_box(odd, floatx, mode, previous, data_format, helpers):
     if data_format == "channels_first" and not len(K._get_available_gpus()):
         return
 
@@ -696,8 +688,8 @@ def test_Backward_Reshape_multiD_box(odd, floatx, mode, previous, data_format):
         decimal = 2
 
     layer = DecomonReshape((-1,), dc_decomp=False, mode=mode, dtype=K.floatx())
-    inputs = get_tensor_decomposition_multid_box(odd, dc_decomp=False)
-    inputs_ = get_standard_values_multid_box(odd, dc_decomp=False)
+    inputs = helpers.get_tensor_decomposition_multid_box(odd, dc_decomp=False)
+    inputs_ = helpers.get_standard_values_multid_box(odd, dc_decomp=False)
     x, y, z_, u_c, W_u, b_u, l_c, W_l, b_l = inputs_
 
     if mode == "hybrid":
@@ -742,7 +734,7 @@ def test_Backward_Reshape_multiD_box(odd, floatx, mode, previous, data_format):
         output_ = f_dense(inputs_)
     w_u_, b_u_, w_l_, b_l_ = output_
 
-    assert_output_properties_box_linear(
+    helpers.assert_output_properties_box_linear(
         x,
         None,
         z_[:, 0],
